fixed bugs

+++ src/app/components/user-management/delete/user-delete.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from "@angular/forms";
-import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
+import {Component, OnInit} from '@angular/core';
+import {FormControl, FormGroup} from "@angular/forms";
+import {Router} from "@angular/router";
+import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 
 import { AdminService } from "../../../services/admin.service";
+import { AuthService } from "../../../services/auth.service";
 import { UserService } from "../../../services/user.service";
 
 import { User } from "../../../types/user";
@@ -42,11 +44,13 @@ export class UserDeleteComponent implements OnInit {
   /**
    * Constructor
    * @constructor
-   * @param {AdminService} adminService service providing admin functionalities
+   * @param {Router} router router providing navigation
+   * @param {AuthService} authService service providing authentication functionalities
    * @param {UserService} userService service providing user functionalities
+   * @param {AdminService} adminService service providing admin functionalities
    * @param {NgbActiveModal} activeModal modal containing this component
    */
-  constructor(public userService : UserService, public adminService: AdminService, public activeModal: NgbActiveModal) {
+  constructor(public router: Router, public authService: AuthService, public userService: UserService, public adminService: AdminService, public activeModal: NgbActiveModal) {
     this.userDeleteForm.disable();
   }
 
@@ -76,6 +80,21 @@ export class UserDeleteComponent implements OnInit {
           console.error('There was an error!', error);
         }
       })
+    } else {
+      this.userService.getUserDetails().subscribe({
+        next: res => {
+          this.user = res;
+
+          this.userDeleteForm.controls['firstName'].setValue(res.firstName);
+          this.userDeleteForm.controls['lastName'].setValue(res.lastName);
+          this.userDeleteForm.controls['email'].setValue(res.email);
+          this.userDeleteForm.controls['role'].setValue(res.role);
+          this.userDeleteForm.controls['notificationChannel'].setValue(res.notificationChannel);
+        },
+        error: error => {
+          console.error('There was an error!', error);
+        }
+      })
     }
   }
 
@@ -83,7 +102,7 @@ export class UserDeleteComponent implements OnInit {
    * Deletes user
    */
   public async deleteUser(): Promise<void> {
-    if (this.user.id != null) {
+    if (this.user.id != null && this.user.id != this.authService.getUserId()) {
       this.adminService.deleteUser(this.user.id).subscribe({
         next: () => {
           this.activeModal.close('deleted');
@@ -93,14 +112,17 @@ export class UserDeleteComponent implements OnInit {
         }
       });
       return;
+    } else {
+      this.userService.deleteUser().subscribe({
+        next: () => {
+          this.activeModal.close('deleted');
+          this.router.navigateByUrl("/");
+          localStorage.clear();
+        },
+        error: error => {
+          console.error('There was an error!', error);
+        }
+      });
     }
-    this.userService.deleteUser().subscribe({
-      next: () => {
-        this.activeModal.close('deleted');
-      },
-      error: error => {
-        console.error('There was an error!', error);
-      }
-    });
   }
 }
